% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_url.R
\name{append_url}
\alias{append_url}
\title{append to URL for API data request}
\usage{
append_url(variable_name, req_url = req_url)
}
\arguments{
\item{variable_name}{name of the filter variable}

\item{req_url}{the url to be appended to}
}
\description{
append to URL for API data request
}
