% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_historical_lmps_national.R
\name{calculate_expected_rows}
\alias{calculate_expected_rows}
\title{pull_historical_lmps_national}
\usage{
calculate_expected_rows(start_date, end_date, num_locations)
}
\arguments{
\item{start_date}{The first day of data to pull}

\item{end_date}{NOT INCLUSIVE! Gridstatus returns data for dates BEFORE this date,
but does not return data for this date}
}
\value{
a table of historical hourly LMPs for all ISOs (wide format). The function
also saves an excel file with all the data
}
\description{
Pulls Historical DA LMPs from Gridstatus for all ISOs for a given date range
}
