% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_query_url.R
\name{construct_query_url}
\alias{construct_query_url}
\title{Build URL for API data request}
\usage{
construct_query_url(
  wh_dataset,
  start_time = Sys.Date() - 5,
  end_time = Sys.Date(),
  limit = 100,
  timezone = "market",
  location = "",
  location_type = "",
  resample_frequency = "",
  respondent = ""
)
}
\arguments{
\item{wh_dataset}{Name of dataset to download}

\item{start_time}{Start time. Default is 5 days ago.}

\item{end_time}{End time. Default is today's date.}

\item{limit}{Number of requests to return. Default is 100.}

\item{timezone}{Default is market, which returns times in the market's local time}

\item{location}{filter term for the data's location column}

\item{location_type}{filter term for the data's location_type column}

\item{resample_frequency}{resample the data to a lower frequency}

\item{respondent}{for specifying EIA fuel mix zone}
}
\value{
req_url URL for API request with specified parameters
}
\description{
Build URL for API data request
}
\examples{
req_url <- construct_query_url("caiso_fuel_mix", "2024-04-1", "2024-04-10")


}
\seealso{
\code{\link[=get_available_datasets]{get_available_datasets()}}
}
