% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_da_hourly_prices.R
\name{get_da_hourly_prices}
\alias{get_da_hourly_prices}
\title{Easily gather hourly DA prices}
\usage{
get_da_hourly_prices(
  iso,
  start_time = Sys.Date() - 5,
  end_time = Sys.Date(),
  limit = 100,
  timezone = "market",
  location = "",
  location_type = ""
)
}
\arguments{
\item{iso}{Name of dataset to download}

\item{start_time}{Start time. Default is 5 days ago.}

\item{end_time}{End time. Default is today's date.}

\item{limit}{Number of requests to return. Default is 100.}

\item{timezone}{Default is market, which returns times in the market's local time}

\item{location}{filter term for the data's location column}

\item{location_type}{filter term for the data's location_type column}
}
\value{
dataset of hourly DA Prices for the specified ISO and date range
}
\description{
Easily gather hourly DA prices
}
\seealso{
\code{\link[=construct_query_url(), get_api_request]{construct_query_url(), get_api_request()}}
}
