% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_load_iso.R
\name{get_load_iso}
\alias{get_load_iso}
\title{Easily gather hourly DA prices}
\usage{
get_load_iso(
  iso,
  start_time = Sys.Date() - 5,
  end_time = Sys.Date(),
  limit = 100,
  timezone = "market",
  location = "",
  location_type = "",
  resample_frequency = "",
  respondent = "",
  tac_area_name = ""
)
}
\arguments{
\item{iso}{Name of dataset to download}

\item{start_time}{Start time. Default is 5 days ago.}

\item{end_time}{End time. Default is today's date.}

\item{limit}{Number of requests to return. Default is 100.}

\item{timezone}{Default is market, which returns times in the market's local time}

\item{location}{filter term for the data's location column}

\item{location_type}{filter term for the data's location_type column}

\item{resample_frequency}{resample the data to a lower frequency}

\item{respondent}{for specifying EIA fuel mix zone}

\item{tac_area_name}{for specifying CAISO Zonal load zone (any zone in WECC)}
}
\value{
dataset of hourly DA Prices for the specified ISO and date range
}
\description{
Easily gather hourly DA prices
}
