% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgridstatus2-package.R
\docType{package}
\name{rgridstatus2-package}
\alias{rgridstatus2}
\alias{rgridstatus2-package}
\title{rgridstatus2: API Wrapper For GridStatus}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bpfeffer/rgridstatus2}
  \item Report bugs at \url{https://github.com/bpfeffer/rgridstatus2/issues}
}

}
\author{
\strong{Maintainer}: Ben Pfeffer \email{bpfffr@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
